benches: 调整bench的目录结构
